Add (most of) the bar plot examples from the docs
This isn't all of them, and isn't a one-for-one copy of the docs. The main
reason being that it looks like the released version of Plotext (2022-10-30)
is quite a bit behind `main` of their repo, and the docs are within the
repo, and there's been changes in the past year.

So for now there's a few things commented out.