Switch to plan B -- pull a figure out from the internals
There's still more work to do on this, but the initial approach was rendered
useless by some of the finer issues of Plotext having global state and
side-effects. Given that it all seems to be working on a single global
"figure" instance, here I work to use a figure per widget to handle things.

This approach does mean that there will need to be a `plot` protocol for any
plots; that is: any plot a developer wishes to do will need to be done in a
subclass of the plot class *and* will need to be implemented in a `plot`
method.

To keep things similar to what they're going to be used to (if they're used
to plotext) a `self.plt` property is provided.

Not everything that the plotext module provides is provided on the figure,
and I've not taken that into account just yet; there's more to follow here.
I just wanted to push this approach as far as possible and see if I could
get the demo working to the point where it broke.

I've got to the point in the demo where it broke and now it works. This
feels like the right direction.