Stub out save_fig
It could make sense to support this in the future, perhaps, but there's a
danger people will copy over their scripts and start spitting out files all
over the place when they don't intend to. There's also the problem of how
this would work with textual-web and where the file should end up.

So, for now, stub it out. We can add it in later.