Note that just GIF plots aren't supported
We can and should support image plots, assuming that the developer has
Pillow in their environment. This is falling back on Plotext itself and in
testing it works fine.

GIF plots won't work out of the box because they're a tight loop and would
be a bad thing to use. I think they can be made to work though but need some
extra code.