Remove the need for the PlotextPlot.plot method
Requested by @willmcgugan -- this commit removes the need for the
PlotextPlot.plot method; the practical impact of this is that:

- The plotting environment isn't a fresh one each render.
- This requires that plotext itself behaves in a predictable way each build() call.
- This potentially makes it easier to decide how and where to build the plot.
- This potentially allows plotting without subclassing.

One downside discovered so far is that Plotext can't be trusted when it
comes to logarithmic scales. For example, even in the repl, without this
library involved, this happens:

>>> import plotext as plt
>>> plt.xscale("log")
>>> plt.plot(plt.sin(periods=2, length=10**4))
>>> plt.show()
<plot is drawn in the terminal here>
>>> plt.show()
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/davep/develop/python/textual-plotext/.venv/lib/python3.7/site-packages/plotext/_core.py", line 293, in show
    _figure.show()
  File "/Users/davep/develop/python/textual-plotext/.venv/lib/python3.7/site-packages/plotext/_figure.py", line 319, in show
    self.build()
  File "/Users/davep/develop/python/textual-plotext/.venv/lib/python3.7/site-packages/plotext/_figure.py", line 326, in build
    self._build_matrix()
  File "/Users/davep/develop/python/textual-plotext/.venv/lib/python3.7/site-packages/plotext/_figure.py", line 332, in _build_matrix
    self.monitor.build_plot() if not self.monitor.fast_plot else None
  File "/Users/davep/develop/python/textual-plotext/.venv/lib/python3.7/site-packages/plotext/_build.py", line 37, in build_plot
    self.x = [ut.apply_scale(self.x[s], xscale[s] is self.default.xscale[1]) for s in Signals] # apply optional log scale
  File "/Users/davep/develop/python/textual-plotext/.venv/lib/python3.7/site-packages/plotext/_build.py", line 37, in <listcomp>
    self.x = [ut.apply_scale(self.x[s], xscale[s] is self.default.xscale[1]) for s in Signals] # apply optional log scale
  File "/Users/davep/develop/python/textual-plotext/.venv/lib/python3.7/site-packages/plotext/_utility.py", line 514, in apply_scale
    return log(data) if test else data
  File "/Users/davep/develop/python/textual-plotext/.venv/lib/python3.7/site-packages/plotext/_utility.py", line 62, in log
    return [math.log10(el) for el in data] if isinstance(data, list) else math.log10(data)
  File "/Users/davep/develop/python/textual-plotext/.venv/lib/python3.7/site-packages/plotext/_utility.py", line 62, in <listcomp>
    return [math.log10(el) for el in data] if isinstance(data, list) else math.log10(data)
ValueError: math domain error

This is the only instance of seen of this I've seen in the README examples
I've tried. However, it does mean that the logarithmic example has to be
left out of the demo for now.

Also: loading up the plots in `on_mount` has the undesirable effect of doing
all the work up front, so that means the demo takes a bit longer to load --
before, because it was done in render, it meant that the plots were drawn up
as needed.

Note that lots of followup work will need to take place after this commit.
All the documentation I've written will need a revisit -- README and
docstrings -- and there's probably some tidying up of the code to take place
too. This commit simply draws in the core change to test this approach out.