Change the default light and dark mode themes to better fit Textual
See #3 and https://github.com/Textualize/textual/issues/3532 for some
background. Some Plotext themes use ANSI colours and "hard code" things;
others blend a lot better with Textual. In further testing it seems that out
of the box "clear" and "pro" are a better fit within Textual *and* they work
well with light/dark mode.

So while not the most colourful of themes, they are the most pleasing to the
eye as a default experience. Because of this, let's make Pro the default in
both cases.