Move to auto-switch of light/dark themes off by default
Instead, by default, use a theme that is far more textual-friendly and which
is responsive to light and dark mode to start with.

This means https://github.com/Textualize/textual/issues/3532 is less of an
issue right out of the gate as the default theme will blend in way better.