Add Textual-specific dark and light themes
And, in doing so, make them the default auto-theme themes, and turn
auto-theme back on by default.

Overall this isn't really going to make much of a difference to how a plot
looks out of the box; but it does mean we can tweak things a little as time
goes on. It also means that if we change anything in the Textual design,
this should follow.

Right now, initially, it simply uses $accent for the ticks and leaves the
background as the default.