Add a "swatch mode" to the theme example
I think we may want to dial in the colour sequence (in other words the
colours to use for subsequent plots of data) for some of the themes --
especially the textual-design-* themes -- so it will make life easier if
there's a mode where you can clearly see what colours are used where.

This should be it. It's like a test card mode.