Settle on a hand-curated set of colours for the textual-design themes
This is still up for tweaking, but I feel this is narrowing it down to a
pleasing palette of colours that work for dark and light mode, and for up to
20 plots in a single plot.

I've kept almost all of the plotext-derived themes using their own sets, as
I believe that's what people will be expecting, but for a couple (default
and pro) I go with this curated set this seems like it'll likely be what
people will use out of the box (actually I suspect most people will simply
go with the textual-design-* themes).

This might get a bit more tweaking; but I think this is close and sets the
stage.