Force a resize of the plot, like, really be clear about it
See #5. I'm not 100% clear on why this works, and I think there's a lot of
code that I'd need to chase to really figure out what's going on here.
But... this works and solves the problem of plot sizes capping at their size
when the last clear_figure happened.

This does seem to allow a figure to be rebuilt, without needing to clear it
and recreate it from scratch.